fusion task, tasks, scheduler dans taskmanager todo : tests
