+++ templates/web/admin/tree.html.twig
@@ -64,13 +64,7 @@
 
   {% if user.ACL().has_right('taskmanager') %}
   <li class="{% if feature == 'taskmanager' %}selected{% endif %}">
-    <a target="right" href="taskmanager.php">
-      <img src="/skins/admin/TaskManager.png" />
-      <span>{% trans 'admin::utilisateurs: gestionnaire de taches' %}</span>
-    </a>
-  </li>
-  <li class="{% if feature == 'taskmanager' %}selected{% endif %}">
-    <a target="right" href="/admin/tasks/" class="ajax">
+    <a target="right" href="/admin/task-manager/tasks/" class="ajax">
       <img src="/skins/admin/TaskManager.png" />
       {% trans 'admin::utilisateurs: gestionnaire de taches 2' %}
     </a>
