+++ www/.htaccess
@@ -6,7 +6,7 @@
   RewriteRule ^(mail-export)/([a-zA-Z0-9]*)\/?$   /include/download_anonymous.php?type=$1&token=$2 [L]
   RewriteRule ^(mail-export)/([a-zA-Z0-9]*)\/get$ /include/download_anonymous.php?type=$1&token=$2&get=1 [L]
 
-  RewriteRule ^admin/.*$                /admin/router.php [L]
+  RewriteRule ^admin/$                /admin/router.php [L]
   RewriteRule ^admin/databox/.*$     /admin/router.php [L]
   RewriteRule ^admin/sphinx/.*$       /admin/router.php [L]
   RewriteRule ^admin/structure/.*$    /admin/router.php [L]
@@ -25,6 +25,7 @@
   RewriteRule ^admin/subdefs/.*$      /admin/router.php [L]
   RewriteRule ^admin/tasks/.*$        /admin/router.php [L]
   RewriteRule ^admin/task/.*$         /admin/router.php [L]
+  RewriteRule ^admin/task-manager/.*$    /admin/router.php [L]
   RewriteRule ^admin/scheduler/.*$    /admin/router.php [L]
 
   RewriteRule ^prod/records/edit/.*$  /prod/router.php [L]
